requirements.txt
chardet==3.0.4
clipboard==0.0.4
defusedxml==0.5.0
distro==1.3.0
lxml==4.3.0
numpy==1.16.0
pandas==0.23.4
pdfminer.six==20181108
pikepdf==1.0.5
pycryptodome==3.7.3
pyperclip==1.7.0
python-dateutil==2.7.5
pytz==2018.9
six==1.12.0
sortedcontainers==2.1.0
tabula-py==1.3.1
urllib3==1.24.1
